+++ src/Cli/SubCommands/HelpCommand.ts
@@ -12,6 +12,7 @@ const HelpCommand: Command = {
         console.log(`  ${chalk.yellow('down')}         Stops the Jetpax server`)
         console.log(`  ${chalk.yellow('ui')}           Opens the Jetpax dashboard`)
         console.log(`  ${chalk.yellow('web')}          Opens the Jetpax dashboard in a browser`)
+        console.log(`  ${chalk.yellow('logs')}         Tails the Jetpax server logs`)
         console.log(`  ${chalk.yellow('help')}         Displays a list of commands`)
         console.log()
     }