+++ src/Cli/SubCommands/index.ts
@@ -5,6 +5,7 @@ import Command from "./Command"
 import StartServerCommand from "./StartServerCommand"
 import OpenWebUiCommand from "./OpenWebUiCommand"
 import StopServerCommand from "./StopServerCommand"
+import TailServerLogsCommand from "./TailServerLogsCommand"
 
 const commands: Command[] = [
     GetVersionCommand,
@@ -13,6 +14,7 @@ const commands: Command[] = [
     OpenUiCommand,
     StartServerCommand,
     StopServerCommand,
+    TailServerLogsCommand,
 ]
 
 export default commands