+++ src/Common/serverLogFilePath.ts
@@ -1,0 +1,5 @@
+import jetpaxHome from "./jetpaxHome"
+
+const serverLogFilePath = `${jetpaxHome}/logs/server/server.log`
+
+export default serverLogFilePath