+++ src/Server/Logging/makeLogger.ts
@@ -2,22 +2,21 @@ import pino, {Logger} from 'pino'
 import jetpaxHome from "../../Common/jetpaxHome"
 import fs from "fs"
 import {execSync} from "child_process"
+import serverLogFilePath from "../../Common/serverLogFilePath"
 
 async function makeFileLogger() {
-    const logFilePath = `${jetpaxHome}/logs/server/server.log`
-
     // Start the log file if it doesn't already exists
-    if (!fs.existsSync(logFilePath)) {
+    if (!fs.existsSync(serverLogFilePath)) {
 
         console.log('log file not found, creating...')
 
         // Make sure the directory exists first too, to be safe
         execSync(`mkdir -p ${jetpaxHome}/logs/server`)
 
-        fs.writeFileSync(logFilePath, '')
+        fs.writeFileSync(serverLogFilePath, '')
     }
 
-    return pino({}, pino.destination(logFilePath))
+    return pino({}, pino.destination(serverLogFilePath))
 }
 
 export async function makeLogger(): Promise<Logger> {